chore: prepare cryptoassets future rework by deprecating many functions
