+++ .changeset/chilly-toys-rule.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/cryptoassets": minor
+"@ledgerhq/types-live": minor
+---
+
+Deprecate cryptoassets function to anticipate Removal CAL initiative.
