+++ libs/ledgerjs/packages/cryptoassets/package.json
@@ -25,7 +25,6 @@
     "prewatch": "pnpm build",
     "watch": "tsc --watch",
     "watch:es": "tsc --watch -m esnext --moduleResolution bundler --outDir lib-es",
-    "doc": "documentation readme src/** --section=API --pe ts --re ts --re d.ts",
     "lint": "eslint ./src --no-error-on-unmatched-pattern --ext .ts,.tsx --cache",
     "lint:fix": "pnpm lint --fix",
     "test": "jest",
@@ -45,7 +44,6 @@
     "@types/bs58check": "^2.1.0",
     "@types/jest": "^29.5.10",
     "@types/node": "^22.10.10",
-    "documentation": "14.0.2",
     "jest": "^29.7.0",
     "rimraf": "^4.4.1",
     "source-map-support": "^0.5.21",
